# backend/docs/swagger.yaml
basePath: /
definitions:
  dto.AncestorsResponse:
    properties:
      asset_id:
        example: SWP-001-1-1
        type: string
      path:
        items:
          $ref: '#/definitions/dto.AssetSummary'
        type: array
    required:
    - asset_id
    - path
    type: object
  dto.AssetDetail:
    properties:
      asset_id:
        example: TX-001-1
        type: string
      asset_name:
        example: Transformer 1
        type: string
      asset_type:
        example: TRANSFORMER
        type: string
      commissioned_date:
        description: CommissionedDate is formatted YYYY-MM-DD.
        example: "2019-04-30"
        type: string
        x-nullable: true
      created_at:
        example: "2026-09-20T08:15:00Z"
        format: date-time
        type: string
      manufacturer:
        example: ABB
        type: string
        x-nullable: true
      model:
        example: RESIBLOC
        type: string
        x-nullable: true
      operational_status:
        example: IN_SERVICE
        type: string
      parent_asset_id:
        example: SUB-001
        type: string
        x-nullable: true
      rating_kva:
        example: 1000
        type: number
        x-nullable: true
      serial_number:
        example: SN-99812
        type: string
        x-nullable: true
      updated_at:
        example: "2026-09-20T08:15:00Z"
        format: date-time
        type: string
      voltage_kv:
        example: 22
        type: number
        x-nullable: true
    required:
    - asset_id
    - asset_name
    - asset_type
    - commissioned_date
    - created_at
    - manufacturer
    - model
    - operational_status
    - parent_asset_id
    - rating_kva
    - serial_number
    - updated_at
    - voltage_kv
    type: object
  dto.AssetListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AssetSummary'
        type: array
      limit:
        example: 25
        type: integer
      offset:
        example: 0
        type: integer
      total:
        description: Total is the number of assets matching the filters, across all
          pages.
        example: 206
        type: integer
    required:
    - items
    - limit
    - offset
    - total
    type: object
  dto.AssetNode:
    properties:
      asset_id:
        example: SWB-001-1
        type: string
      asset_name:
        example: Switchboard 1
        type: string
      asset_type:
        example: SWITCHBOARD
        type: string
      child_count:
        description: ChildCount is the number of immediate children (0 = leaf).
        example: 4
        type: integer
      commissioned_date:
        example: "2019-04-30"
        type: string
        x-nullable: true
      operational_status:
        example: IN_SERVICE
        type: string
      parent_asset_id:
        example: SUB-001
        type: string
        x-nullable: true
      rating_kva:
        description: |-
          RatingKVA and CommissionedDate let a list of nodes (the children of an
          asset) be sorted without a request per row. CommissionedDate is YYYY-MM-DD.
        example: 1000
        type: number
        x-nullable: true
      subtree_count:
        description: |-
          SubtreeCount is the number of assets in this asset's subtree,
          including the asset itself (a leaf has 1).
        example: 5
        type: integer
    required:
    - asset_id
    - asset_name
    - asset_type
    - child_count
    - commissioned_date
    - operational_status
    - parent_asset_id
    - rating_kva
    - subtree_count
    type: object
  dto.AssetSummary:
    properties:
      asset_id:
        example: TX-001-1
        type: string
      asset_name:
        example: Transformer 1
        type: string
      asset_type:
        example: TRANSFORMER
        type: string
      operational_status:
        example: IN_SERVICE
        type: string
      parent_asset_id:
        example: SUB-001
        type: string
        x-nullable: true
    required:
    - asset_id
    - asset_name
    - asset_type
    - operational_status
    - parent_asset_id
    type: object
  dto.ChildGroup:
    properties:
      asset_type:
        example: SWITCHBOARD
        type: string
      assets:
        items:
          $ref: '#/definitions/dto.AssetNode'
        type: array
      count:
        description: Count is the number of immediate children of this type.
        example: 2
        type: integer
      subtree_count:
        description: |-
          SubtreeCount is the combined subtree size of those children
          (each child counted with its own descendants).
        example: 9
        type: integer
    required:
    - asset_type
    - assets
    - count
    - subtree_count
    type: object
  dto.ChildrenResponse:
    properties:
      asset_id:
        example: SUB-001
        type: string
      descendant_counts:
        items:
          $ref: '#/definitions/dto.TypeCount'
        type: array
      groups:
        items:
          $ref: '#/definitions/dto.ChildGroup'
        type: array
      total_children:
        example: 6
        type: integer
    required:
    - asset_id
    - descendant_counts
    - groups
    - total_children
    type: object
  dto.ErrorResponse:
    properties:
      error:
        description: |-
          Error is a human-readable, user-safe message. For 5xx it is generic —
          internals are only ever written to the server log.
        example: file "assets.txt" must have a .csv extension
        type: string
      line:
        description: Line is the 1-indexed source line of a structurally malformed
          CSV, when known.
        example: 42
        type: integer
      trace_id:
        description: |-
          TraceID matches the X-Trace-Id response header and the server log lines
          for this request; quote it when reporting a 5xx.
        example: 7f6c1a52-3c1e-4a51-9d0b-5b1b2f7a9c11
        type: string
    required:
    - error
    type: object
...
  dto.RootsResponse:
    properties:
      roots:
        items:
          $ref: '#/definitions/dto.AssetNode'
        type: array
      total:
        example: 13
        type: integer
    required:
    - roots
    - total
    type: object
  dto.SearchResponse:
    properties:
      count:
        example: 2
        type: integer
      query:
        example: tx-001
        type: string
      results:
        items:
          $ref: '#/definitions/dto.AssetSummary'
        type: array
      truncated:
        description: Truncated is true when more assets matched than limit allowed.
        example: false
        type: boolean
      type:
        description: Type echoes the applied asset_type filter ("" = none).
        example: TRANSFORMER
        type: string
    required:
    - count
    - query
    - results
    - truncated
    - type
    type: object
  dto.StatsResponse:
    properties:
      by_status:
        items:
          $ref: '#/definitions/dto.StatusCount'
        type: array
      by_type:
        items:
          $ref: '#/definitions/dto.TypeCount'
        type: array
      total:
        example: 206
        type: integer
    required:
    - by_status
    - by_type
    - total
    type: object
  dto.StatusCount:
    properties:
      count:
        example: 146
        type: integer
      operational_status:
        example: IN_SERVICE
        type: string
    required:
    - count
    - operational_status
    type: object
  dto.TypeCount:
    properties:
      asset_type:
        example: TRANSFORMER
        type: string
      count:
        example: 3
        type: integer
    required:
    - asset_type
    - count
    type: object
info:
  contact: {}
  description: CSV asset import/validation and asset hierarchy explorer API.
  title: Smart Grid Asset Management API
  version: "1.0"
paths:
  /api/assets:
    get:
      description: One page of every stored asset, ordered by asset id unless `sort`
        says otherwise, optionally filtered by text (case-insensitive substring of
        the id or name), asset type and operational status. `total` counts all matches
        across pages.
      parameters:
      - description: Text to look for in the asset id or name
        in: query
        maxLength: 200
        name: q
        type: string
      - description: Only assets of this type (see GET /api/lookups)
        in: query
        name: type
        type: string
      - description: Only assets in this operational status (see GET /api/lookups)
        in: query
        name: status
        type: string
      - default: asset_id
        description: Column to order by; ties and the default are by asset id. Assets
          with no parent sort last.
        enum:
        - asset_id
        - name
        - type
        - status
        - parent
        in: query
        name: sort
        type: string
      - default: asc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      - default: 25
        description: Page size
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Rows to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AssetListResponse'
        "400":
          description: Unknown type, status or sort, or bad dir/limit/offset
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all assets
      tags:
      - assets
  /api/assets/{assetId}:
    delete:
      description: Removes an asset that has no children. An asset that still has
        children is refused with 409 and must have them deleted first. See docs/KNOWN_LIMITATIONS.md.
      parameters:
      - description: Asset id
        example: TX-001-1
        in: path
        name: assetId
        required: true
        type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: No API key is configured on the server, so the API is disabled
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: No such asset
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: The asset has children, or another writer collided with the
            delete
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete one asset
      tags:
      - assets
    get:
      description: Every stored attribute of the asset.
      parameters:
      - description: Asset id
        example: TX-001-1
        in: path
        name: assetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AssetDetail'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: No such asset
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get one asset
      tags:
      - assets
  /api/assets/{assetId}/ancestors:
    get:
      description: The path from the top of the hierarchy down to the asset, in that
        order; the asset itself is the last element. Use it to expand a tree to a
        search hit.
      parameters:
      - description: Asset id
        example: SWP-001-1-1
        in: path
        name: assetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AncestorsResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: No such asset
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get an asset's ancestor path
      tags:
      - assets
  /api/assets/{assetId}/children:
    get:
      description: 'The asset''s immediate children grouped by asset type (each with
        `child_count` and `subtree_count`), plus `descendant_counts`: how many descendants
        of each type sit anywhere beneath the asset. An asset with no children returns
        empty lists.'
      parameters:
      - description: Asset id
        example: SUB-001
        in: path
        name: assetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ChildrenResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: No such asset
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List an asset's children
      tags:
      - assets
  /api/assets/roots:
    get:
      description: Every asset that has no parent (substations), each with `child_count`
        and `subtree_count` so a tree can render expanders without further requests.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RootsResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List top-level assets
      tags:
      - assets
  /api/assets/search:
    get:
      description: Case-insensitive substring match on asset id or name, optionally
        restricted to one asset type. Exact id matches come first, then id prefixes.
        `%` and `_` in `q` are matched literally. Results are capped at `limit`; `truncated`
        says whether more matched.
      parameters:
      - description: Text to look for in the asset id or name
        in: query
        maxLength: 200
        minLength: 1
        name: q
        required: true
        type: string
      - description: Only assets of this type (see GET /api/lookups)
        in: query
        name: type
        type: string
      - default: 50
        description: Maximum results
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SearchResponse'
        "400":
          description: Missing q, unknown type, or bad limit
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search assets
      tags:
      - assets
  /api/assets/stats:
    get:
      description: The total number of stored assets and how many there are of each
        asset type and each operational status. Types and statuses with no assets
        are omitted.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StatsResponse'
        "401":
          description: Missing or invalid x-api-key
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Count assets by type and status
      tags:
      - assets
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: API key (see API_KEY in .env). Required by every /api endpoint (sent
      as the x-api-key header). Only /healthz and these docs are open. This is a shared
      key, not user authentication (see ARCHITECTURE.md).
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"